Group math operations, move randpy
