+++ analyze/checkerflickerplus.py
@@ -12,7 +12,7 @@ from pathlib import PureWindowsPath
 import os
 import warnings
 
-from ..external import randpy
+from .. import randpy
 from ..modules import analysisfuncs as asc
 from ..modules import iofuncs as iof
 from ..modules import miscfuncs as msc
