+++ analyze/omb.py
@@ -10,7 +10,7 @@ import numpy as np
 import os
 from scipy.stats.mstats import mquantiles
 
-from ..external import randpy
+from .. import randpy
 from ..modules import analysisfuncs as asc
 from ..modules import iofuncs as iof
 from ..plot import util as plf
