+++ calc/__init__.py
@@ -1,0 +1,18 @@
+"""
+Mathematical operations for pymer.
+"""
+from . import gaussfitter
+from .primefac import (
+    isprime,
+    prime_sieve,
+    factorize,
+    prime_factorize,
+)
+
+__all__ = [
+    'gaussfitter',
+    'isprime',
+    'prime_sieve',
+    'factorize',
+    'prime_factorize',
+]
