+++ calc/gaussfitter.py
