--- external/__init__.py
@@ -1,7 +1,0 @@
-from . import gaussfitter
-from .randpy import randpy
-
-__all__ = [
-    'gaussfitter',
-    'randpy',
-]
