+++ randpy/rng_gasdev_ran1.cpp
